ADD Check playerNameAvailability on register
-Add an endpoint to check Name Availability
(.../player/isNameAvailable/{playerName}) to be used in the registration
process. It can be used to check if a name is available or not so you
can warn the new User to change it before doing the POST of newUser
(before clicking Register).

-Add a httpResponse to register newUser, to confirm the user was added
successfully or it failed because the playerName is not available.