Statics, PlayerDTOs
-Changes in DTOs for players. Decided not to use ModelMapper in players
because i want to add attributes to objects(Stats) so the model mapper
is not usefull (for me) to add mappings if the the attribute is a
Transient in the entity. The conversion was made by a new class
(playerDTO) and mapped via constructor.

-Changes in Statics. Moved part of the logic to the service. Add if in
constructor to avoid math error when dived by 0.
-IMPORTANT. To implement and use the StaticsService, that is not linked
with an Entity (Statics are not persisted and constructed volatile) a
Bean is need to invoke the StaticsService when Autowire in Player or
games Controller.


TO DO:
-Delete DTO endpoints ()were used for construction process)
-Ranking for the best and the worst