JDBC WORKING
-PLAYER
--Return @Transient annotation. Note constructor full of arguments for
jdbc repository marked with @PersistanceConstructor and an empty
constructor for RowMapper( not in used yet for player!)
--DTO conversion done with a manual class PlayerDTO. This DTO includes
stats calculation (stats not persisted).

-STATS
--Working for AvgIsWin Calculation
--Stats Controller working for all games and for a player
TO DO: Ranking Best and Worst

-GAMES
--Game Service tuned.
--**Delete method is no longer needded since the game table is build
with cascade property, so when deleting a player the RDBMS does
associated deletion (all games for that player)
--Games DTO is no longer used since there is no more @Entity JPA
annotation. This means that in Games there is no more Player Object
inside each Game. There is only a playerId persisted and related in DB
via foreign Key. I keep it for future DTO Game conversion if needed

-GENERAL
--RowMappers not used yet. The @Query construction in the repository
working with a manual DTO conversion is working for now.
--RowMapper used in queries executed by jdbcTemplate. Example:
jdbcTemplate.queryForList or ...ForObject, etc. In the arguments needs a
rowMapper